fix: the two reds main already carried — the teardown ledger moves to its own module, and the rsp CI job installs the binary its residents need
Two failures predate the union-repair branch and block the same Version PR:

- apps/redskilled/src/daemon/lifecycle.ts crossed its shrink-only file-size
  baseline (2523 > 2489) when #3752 introduced the exactly-once teardown
  ledger inline. The ledger — the workerStops map and the endWorkerOnce
  closure — moves to worker-teardown.ts, behavior byte-equivalent, and
  lifecycle.ts lands at 2484.

- The rsp CI job (#3625, added today, never green) copied
  REDDB_SKIP_POSTINSTALL: "1" from the dev-oriented install steps, skipping
  the @reddb-io/sdk `red` binary its resident tests genuinely open. The rsp
  job now lets the sdk postinstall run; every other job keeps the skip.

Refs #3625 #3752
